+++ crates/umem_grpc_server/build.rs
@@ -3,8 +3,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .build_server(true)
         .build_client(false)
         .type_attribute("Memory", "#[derive(serde::Serialize, serde::Deserialize)]")
+        // .out_dir("src")
         .type_attribute(
-            "ModifyMemoryParameters",
+            "UpdateMemoryParameters",
             "#[derive(serde::Serialize, serde::Deserialize)]",
         )
         .compile_protos(&["proto/memory.proto"], &["proto"])?;
