+++ crates/umem_grpc_server/proto/memory.proto
@@ -8,31 +8,30 @@ service MemoryService {
   // Qdrant Mutations
   rpc AddMemory(Memory) returns (google.protobuf.Empty);
   rpc AddMemoryBulk(MemoryBulk) returns (google.protobuf.Empty);
-  rpc ModifyMemory(ModifyMemoryParameters) returns (google.protobuf.Empty);
+  rpc UpdateMemory(UpdateMemoryParameters) returns (google.protobuf.Empty);
   rpc DeleteMemory(DeleteMemoryParameters) returns (google.protobuf.Empty);
 
   // Qdrant Queries
-  rpc GetMemoryByQuery(GetMemoriesByQueryParameters) returns (MemoryBulk);
-  rpc GetMemoryByUserID(GetMemoriesByUserIDParameters) returns (MemoryBulk);
+  rpc GetMemoriesByQuery(GetMemoriesByQueryParameters) returns (MemoryBulk);
+  rpc GetMemoriesByUserID(GetMemoriesByUserIDParameters) returns (MemoryBulk);
 }
 
 message Memory {
   string user_id = 1; // tenant
   string memory_id = 2;
   string content = 3;
-  string priority = 4;
+  int32 priority = 4;
   repeated string tags = 5;
-
-  string created_at = 6;
-  string updated_at = 7;
+  int64 created_at = 6;
+  int64 updated_at = 7;
 }
 
 message MemoryBulk { repeated Memory memories = 1; }
 
-message ModifyMemoryParameters {
+message UpdateMemoryParameters {
   string memory_id = 1;
   string content = 2;
-  string priority = 3;
+  int32 priority = 3;
   repeated string tags = 4;
 }
