+++ crates/umem_vector/Cargo.toml
@@ -10,3 +10,5 @@ tokio = { workspace = true, features = ["rt-multi-thread"] }
 tonic = "0.13.1"
 uuid = { version = "1.17.0", features = ["v4"] }
 serde_json = {workspace = true}
+serde = {workspace = true}
+chrono = "0.4.41"
